+++ TowerDefence/src/se/chalmers/slickTD/ProjectileView.java
@@ -13,7 +13,7 @@ public class ProjectileView {
 		}
 	}
 	
-	public void draw(int x,int y){
+	public void draw(float x,float y){
 		image.draw(x,y);	
 	}
 }
